Add more compiler flags. Fix problems

+++ src/Sensor/pi_2_dht_read.cpp
@@ -122,7 +122,7 @@ int pi_2_dht_read(int type, int pin, float* humidity, float* temperature)
     // threshold.
     // Ignore the first two readings because they are a constant 80 microsecond
     // pulse.
-    uint32_t threshold = 0;
+    size_t threshold = 0;
     for (int i = 2; i < DHT_PULSES * 2; i += 2) {
         threshold += pulseCounts[i];
     }
